+++ CHANGELOG.md
@@ -7,10 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] -- 0000-00-00
 
+### Added
+
+- Bugfix: Map text rendering not accounting for hi-res flag
+- Hook: TextRenderA
+- Hook: MapRenderB, MapRenderA
+
 ### Fixed
 
 - QuickRace: Cantina not synced when setting number of racers
 - Updater: Mouse cursor not visible when showing restart message
+- Hook: TextRenderB insertion point after text render
 
 ## [0.1.3] -- 2024-04-30